Move summary methods into summary model
- Moved summary methods into their own model
