+++ app/Models/ResourceTypeItemType/AllocatedExpense.php
@@ -246,86 +246,6 @@ class AllocatedExpense extends Model
     }
 
     /**
-     * Return the summary for all items for the resources in the requested resource type
-     *
-     * @param int $resource_type_id
-     * @param boolean $include_unpublished = false
-     *
-     * @return array
-     */
-    public function summary(int $resource_type_id, bool $include_unpublished): array
-    {
-        $collection = $this->selectRaw('sum(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join('resource', 'item.resource_id', 'resource.id')->
-            join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
-            where('resource_type.id', '=', $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->
-            get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type grouped by resource
-     *
-     * @param int $resource_type_id
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function resourcesSummary(int $resource_type_id, bool $include_unpublished): array
-    {
-        $collection = $this->selectRaw('
-                resource.id AS id, 
-                resource.name AS `name`, 
-                SUM(item_type_allocated_expense.actualised_total) AS total'
-            )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join('resource', 'item.resource_id', 'resource.id')->
-            join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
-            where('resource_type.id', '=', $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy('resource.id')->
-            orderBy('name')->
-            get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type grouped by year
-     *
-     * @param int $resource_type_id
-     * @param boolean $include_unpublished
-
-     * @return array
-     */
-    public function yearsSummary(int $resource_type_id, bool $include_unpublished): array
-    {
-        $collection = $this->selectRaw("
-                YEAR(item_type_allocated_expense.effective_date) as year,
-                SUM(item_type_allocated_expense.actualised_total) AS total"
-            )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            where("resource_type.id", "=", $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("year")->
-            orderBy("year")->
-            get()->
-            toArray();
-    }
-
-    /**
      * Work out the maximum effective date year for the requested
      * resource type id, defaults to the current year if no data exists
      *
@@ -371,304 +291,6 @@ class AllocatedExpense extends Model
         } else {
             return intval($result->year_limit);
         }
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type grouped by month for the requested year
-     *
-     * @param integer $resource_type_id
-     * @param integer $year
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function monthsSummary(int $resource_type_id, int $year, bool $include_unpublished): array
-    {
-        $collection = $this->selectRaw("
-                MONTH(item_type_allocated_expense.effective_date) as month, 
-                SUM(item_type_allocated_expense.actualised_total) AS total"
-            )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            where("resource_type.id", "=", $resource_type_id)->
-            where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $year);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("month")->
-            orderBy("month")->
-            get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type for a specific year and month
-     *
-     * @param integer $resource_type_id
-     * @param integer $year
-     * @param integer $month
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function monthSummary(
-        int $resource_type_id,
-        int $year,
-        int $month,
-        bool $include_unpublished
-    ): array
-    {
-        $collection = $this->selectRaw("
-                MONTH(item_type_allocated_expense.effective_date) as month, 
-                SUM(item_type_allocated_expense.actualised_total) AS total"
-            )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            where("resource_type.id", "=", $resource_type_id)->
-            where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $year)->
-            where(DB::raw('MONTH(item_type_allocated_expense.effective_date)'), '=', $month);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("month")->
-            orderBy("month")->
-            get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type for a specific year
-     *
-     * @param integer $resource_type_id
-     * @param integer $year
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function yearSummary(int $resource_type_id, int $year, bool $include_unpublished): array
-    {
-        $collection = $this->selectRaw("
-                YEAR(item_type_allocated_expense.effective_date) as year, 
-                SUM(item_type_allocated_expense.actualised_total) AS total"
-            )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            where("resource_type.id", "=", $resource_type_id)->
-            where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $year);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("year")->
-            orderBy("year")->
-            get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type grouped by category
-     *
-     * @param integer $resource_type_id
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function categoriesSummary(
-        int $resource_type_id,
-        bool $include_unpublished = false
-    ): array
-    {
-        $collection = $this->selectRaw('
-                category.id, 
-                category.name AS name, 
-                category.description AS description,
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            join("item_category", "item_category.item_id", "item.id")->
-            join("category", "category.id", "item_category.category_id")->
-            where("category.resource_type_id", "=", $resource_type_id)->
-            where("resource_type.id", "=", $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("category.id")->
-            orderBy("name")->
-            get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type for the requested category
-     *
-     * @param integer $resource_type_id
-     * @param integer $category_id
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function categorySummary(
-        int $resource_type_id,
-        int $category_id,
-        bool $include_unpublished = false
-    ): array
-    {
-        $collection = $this->selectRaw('
-                category.id, 
-                category.name AS name, 
-                category.description, 
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            join("item_category", "item_category.item_id", "item.id")->
-            join("category", "category.id", "item_category.category_id")->
-            where("category.resource_type_id", "=", $resource_type_id)->
-            where("resource_type.id", "=", $resource_type_id)->
-            where("category.id", '=', $category_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("category.id")->
-            orderBy("name")->
-            get()->
-            toArray();
-    }
-
-    public function filteredSummary(
-        int $resource_type_id,
-        int $category_id = null,
-        int $subcategory_id = null,
-        int $year = null,
-        int $month = null,
-        array $search_parameters = [],
-        bool $include_unpublished = false
-    ): array
-    {
-        $collection = $this->
-            selectRaw('SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            join("item_category", "item_category.item_id", "item.id")->
-            join("item_sub_category", "item_sub_category.item_category_id", "item_category.id")->
-            join("category", "category.id", "item_category.category_id")->
-            join("sub_category", "sub_category.id", "item_sub_category.sub_category_id")->
-            where("resource_type.id", "=", $resource_type_id);
-
-        if ($category_id !== null) {
-            $collection->where("category.id", "=", $category_id);
-        }
-        if ($subcategory_id !== null) {
-            $collection->where("sub_category.id", "=", $subcategory_id);
-        }
-        if ($year !== null) {
-            $collection->whereRaw(DB::raw("YEAR(item_type_allocated_expense.effective_date) = {$year}"));
-        }
-        if ($month !== null) {
-            $collection->whereRaw(DB::raw("MONTH(item_type_allocated_expense.effective_date) = {$month}"));
-        }
-        if (count($search_parameters) > 0) {
-            foreach ($search_parameters as $field => $search_term) {
-                $collection->where('item_type_allocated_expense.' . $field, 'LIKE', '%' . $search_term . '%');
-            }
-        }
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type and category grouped by subcategory
-     *
-     * @param int $resource_type_id
-     * @param int $category_id
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function subcategoriesSummary(
-        int $resource_type_id,
-        int $category_id,
-        bool $include_unpublished = false
-    ): array
-    {
-        $collection = $this->selectRaw('
-                sub_category.id, 
-                sub_category.name AS name, 
-                sub_category.description AS description, 
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            join("item_category", "item_category.item_id", "item.id")->
-            join("item_sub_category", "item_sub_category.item_category_id", "item_category.id")->
-            join("category", "category.id", "item_category.category_id")->
-            join("sub_category", "sub_category.id", "item_sub_category.sub_category_id")->
-            where("category.resource_type_id", "=", $resource_type_id)->
-            where("resource_type.id", "=", $resource_type_id)->
-            where("category.id", "=", $category_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("sub_category.id")->
-            orderBy("name")->
-            get()->
-            toArray();
-    }
-
-    /**
-     * Return the summary for all items for the resources in the requested resource
-     * type and category and subcategory
-     *
-     * @param int $resource_type_id
-     * @param int $category_id
-     * @param int $subcategory_id
-     * @param boolean $include_unpublished
-     *
-     * @return array
-     */
-    public function subcategorySummary(
-        int $resource_type_id,
-        int $category_id,
-        int $subcategory_id,
-        bool $include_unpublished = false
-    ): array
-    {
-        $collection = $this->selectRaw('
-                sub_category.id, 
-                sub_category.name AS name, 
-                sub_category.description AS description, 
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
-            join("resource", "resource.id", "item.resource_id")->
-            join("resource_type", "resource_type.id", "resource.resource_type_id")->
-            join("item_category", "item_category.item_id", "item.id")->
-            join("item_sub_category", "item_sub_category.item_category_id", "item_category.id")->
-            join("category", "category.id", "item_category.category_id")->
-            join("sub_category", "sub_category.id", "item_sub_category.sub_category_id")->
-            where("category.resource_type_id", "=", $resource_type_id)->
-            where("resource_type.id", "=", $resource_type_id)->
-            where("category.id", "=", $category_id)->
-            where('sub_category.id', '=', $subcategory_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
-        return $collection->groupBy("sub_category.id")->
-            orderBy("name")->
-            get()->
-            toArray();
     }
 
     /**
